# openapi/guzo-v1.json
{
  "components": {
    "schemas": {
...
      "Currency": {
        "enum": [
          "ETB",
          "USD"
        ],
        "title": "Currency",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Money": {
        "description": "An amount in minor units (santim, cents). Never a float.",
        "properties": {
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          }
        },
        "required": [
          "amount_minor",
          "currency"
        ],
        "title": "Money",
        "type": "object"
      },
...
      "PricingStrategyName": {
        "enum": [
          "zone_fixed",
          "per_seat",
          "hourly",
          "quoted",
          "offer"
        ],
        "title": "PricingStrategyName",
        "type": "string"
      },
      "ProductPolicy": {
        "description": "Per-product rules. Bookings keep a snapshot, so edits only affect new bookings.",
        "properties": {
          "free_cancel_hours": {
            "minimum": 0.0,
            "title": "Free Cancel Hours",
            "type": "integer"
          },
          "free_wait_minutes": {
            "minimum": 0.0,
            "title": "Free Wait Minutes",
            "type": "integer"
          },
          "late_cancel_fee_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Late Cancel Fee Pct",
            "type": "integer"
          },
          "no_show_fee_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "No Show Fee Pct",
            "type": "integer"
          }
        },
        "required": [
          "free_cancel_hours",
          "free_wait_minutes",
          "no_show_fee_pct",
          "late_cancel_fee_pct"
        ],
        "title": "ProductPolicy",
        "type": "object"
      },
      "ProductResponse": {
        "properties": {
          "city_id": {
            "title": "City Id",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "policy": {
            "$ref": "#/components/schemas/ProductPolicy"
          },
          "pricing_strategy": {
            "$ref": "#/components/schemas/PricingStrategyName"
          },
          "vehicle_classes": {
            "items": {
              "$ref": "#/components/schemas/VehicleClass"
            },
            "title": "Vehicle Classes",
            "type": "array"
          }
        },
        "required": [
          "code",
          "city_id",
          "name",
          "pricing_strategy",
          "policy",
          "vehicle_classes"
        ],
        "title": "ProductResponse",
        "type": "object"
      },
      "ProductUpdate": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductPolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "vehicle_classes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/VehicleClass"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vehicle Classes"
          }
        },
        "title": "ProductUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VehicleCapacityBody": {
        "properties": {
          "city_id": {
            "default": "addis",
            "title": "City Id",
            "type": "string"
          },
          "max_bags": {
            "maximum": 60.0,
            "minimum": 0.0,
            "title": "Max Bags",
            "type": "integer"
          },
          "max_seats": {
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Max Seats",
            "type": "integer"
          },
          "vehicle_class": {
            "$ref": "#/components/schemas/VehicleClass"
          }
        },
        "required": [
          "vehicle_class",
          "max_seats",
          "max_bags"
        ],
        "title": "VehicleCapacityBody",
        "type": "object"
      },
      "VehicleClass": {
        "enum": [
          "sedan",
          "minivan",
          "suv"
        ],
        "title": "VehicleClass",
        "type": "string"
      },
...
      "ZoneBody": {
        "properties": {
          "city_id": {
            "default": "addis",
            "title": "City Id",
            "type": "string"
          },
          "code": {
            "pattern": "^[a-z0-9_]{2,40}$",
            "title": "Code",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "ZoneBody",
        "type": "object"
      },
      "ZonePriceBody": {
        "properties": {
          "city_id": {
            "default": "addis",
            "title": "City Id",
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          },
          "product_code": {
            "title": "Product Code",
            "type": "string"
          },
          "vehicle_class": {
            "$ref": "#/components/schemas/VehicleClass"
          },
          "zone_id": {
            "title": "Zone Id",
            "type": "string"
          }
        },
        "required": [
          "product_code",
          "zone_id",
          "vehicle_class",
          "price"
        ],
        "title": "ZonePriceBody",
        "type": "object"
      },
      "ZonePriceResponse": {
        "properties": {
          "city_id": {
            "default": "addis",
            "title": "City Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/Money"
          },
          "product_code": {
            "title": "Product Code",
            "type": "string"
          },
          "vehicle_class": {
            "$ref": "#/components/schemas/VehicleClass"
          },
          "zone_id": {
            "title": "Zone Id",
            "type": "string"
          }
        },
        "required": [
          "product_code",
          "zone_id",
          "vehicle_class",
          "price",
          "id"
        ],
        "title": "ZonePriceResponse",
        "type": "object"
      },
      "ZoneResponse": {
        "properties": {
          "city_id": {
            "title": "City Id",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "city_id",
          "name"
        ],
        "title": "ZoneResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Guzo API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/catalog/zones": {
      "get": {
        "operationId": "list_zones",
        "parameters": [
          {
            "in": "query",
            "name": "city_id",
            "required": false,
            "schema": {
              "default": "addis",
              "title": "City Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ZoneResponse"
                  },
                  "title": "Response List Zones",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Zones",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/v1/ops/catalog/products/{code}": {
      "patch": {
        "description": "Change a product's policy or vehicle options. Bookings already made keep their terms.",
        "operationId": "update_product",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "city_id",
            "required": false,
            "schema": {
              "default": "addis",
              "title": "City Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Product",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ops/catalog/vehicle-capacities": {
      "put": {
        "description": "How many riders and bags a vehicle class takes. Quotes refuse anything larger.",
        "operationId": "set_vehicle_capacity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCapacityBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleCapacityBody"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Vehicle Capacity",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ops/catalog/zone-prices": {
      "get": {
        "operationId": "list_zone_prices",
        "parameters": [
          {
            "in": "query",
            "name": "city_id",
            "required": false,
            "schema": {
              "default": "addis",
              "title": "City Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "product_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Product Code"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ZonePriceResponse"
                  },
                  "title": "Response List Zone Prices",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Zone Prices",
        "tags": [
          "ops"
        ]
      },
      "put": {
        "description": "Publish or change the fixed price for a zone and vehicle class.",
        "operationId": "set_zone_price",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZonePriceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZonePriceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Zone Price",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ops/catalog/zone-prices/{price_id}": {
      "delete": {
        "description": "Stop selling a zone and vehicle class. Existing quotes and bookings keep their price.",
        "operationId": "delete_zone_price",
        "parameters": [
          {
            "in": "path",
            "name": "price_id",
            "required": true,
            "schema": {
              "title": "Price Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Zone Price",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/ops/catalog/zones": {
      "put": {
        "operationId": "set_zone",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZoneBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Zone",
        "tags": [
          "ops"
        ]
      }
    },
...
  }
}